backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
numpy
python-dotenv
reportlab
